+++ src/Endpoint/Endpoints/Client.php
@@ -16,4 +16,26 @@ class Client extends AbstractEndpoint
     {
         return $this->client->get('/client');
     }
+    
+    /**
+     * @param string $clientName
+     * 
+     * @return array|null
+     */
+    public function create(string $clientName): ?array
+    {
+        return $this->client->post('/client', [
+            'name' => $clientName
+        ]);
+    }
+    
+    /**
+     * @param int $clientId
+     * 
+     * @return array|null
+     */
+    public function delete(int $clientId): ?array
+    {
+        return $this->client->delete("/client/$clientId");
+    }
 }